Rename black_oil_functions to BlackOilFunctions
